Just use one .c and .h file because I'm bad at Make
