+++ sdl_glue/Makefile
@@ -25,8 +25,8 @@ LIBTARGET = libidrissdl.$(SHLIB_SUFFIX)
 VERSION = 0.1.0
 INSTALLDIR = `${IDRIS} --libdir`/sdl-$(VERSION)
 
-HDRS = idris_sdl.h idris_sdl_ttf.h
-OBJS = idris_sdl.o idris_sdl_ttf.o
+HDRS = idris_sdl.h 
+OBJS = idris_sdl.o
 
 SDL_LIBS := -lSDL2 $(shell sdl2-config --libs) -lSDL2_image -lSDL2_ttf
 SDL_FLAGS := $(shell sdl2-config --cflags)
